+++ components/organisms/footer.tsx
@@ -5,13 +5,10 @@ import { useRouter } from 'next/navigation';
 import AppBar from '@mui/material/AppBar';
 import Toolbar from '@mui/material/Toolbar';
 import Typography from '@mui/material/Typography';
-import { CssBaseline, ThemeProvider, Link, Box } from '@mui/material';
-import { CacheProvider } from '@emotion/react';
-import createCache from '@emotion/cache';
+import { Link, Box } from '@mui/material';
 
 import { subTheme } from '../../styles/theme'
-
-const cache = createCache({ key: 'css', prepend: true });
+import ThemeWrapper from '../../styles/ThemeWrapper';
 
 const Footer = () => {
   const router = useRouter();
@@ -27,31 +24,28 @@ const Footer = () => {
   };
 
   return (
-    <CacheProvider value={cache}>
-      <ThemeProvider theme={subTheme}>
-        <CssBaseline />
-        <AppBar position="fixed" color="secondary" sx={{ top: 'auto', bottom: 0, height: '7vh' }}>
-          <Toolbar>
-            <Box width="100%" display="flex" justifyContent="space-between" sx={{ height: '50%' }}>
-              <Typography variant="body1" color="primary">
-                © 2024 Shimadu Yu
-              </Typography>
-              <Box display="flex" justifyContent="space-between" gap="20px">
-                <Link href="#" color="primary" underline="hover" onClick={goToPreparation}>
-                    利用規約
-                </Link>
-                <Link href="#" color="primary" underline="hover" onClick={goToPreparation2}>
-                    プライバシーポリシー
-                </Link>
-                <Link href="#" color="primary" underline="hover" onClick={goToPreparation3}>
-                    お問い合わせ
-                </Link>
-              </Box>
+    <ThemeWrapper theme={subTheme}>
+      <AppBar position="fixed" color="secondary" sx={{ top: 'auto', bottom: 0, height: '7vh' }}>
+        <Toolbar>
+          <Box width="100%" display="flex" justifyContent="space-between" sx={{ height: '50%' }}>
+            <Typography variant="body1" color="primary">
+              © 2024 Shimadu Yu
+            </Typography>
+            <Box display="flex" justifyContent="space-between" gap="20px">
+              <Link href="#" color="primary" underline="hover" onClick={goToPreparation}>
+                  利用規約
+              </Link>
+              <Link href="#" color="primary" underline="hover" onClick={goToPreparation2}>
+                  プライバシーポリシー
+              </Link>
+              <Link href="#" color="primary" underline="hover" onClick={goToPreparation3}>
+                  お問い合わせ
+              </Link>
             </Box>
-          </Toolbar>
-        </AppBar>
-      </ThemeProvider>
-    </CacheProvider>
+          </Box>
+        </Toolbar>
+      </AppBar>
+    </ThemeWrapper>
   );
 };
