+++ components/organisms/menu.tsx
@@ -1,9 +1,7 @@
 "use client";
 import React, { useState,useEffect } from 'react';
 
-import { Button, CssBaseline, ThemeProvider, Drawer, Box, List, ListItem } from '@mui/material';
-import { CacheProvider } from '@emotion/react';
-import createCache from '@emotion/cache';
+import { Button, Drawer, Box, List, ListItem } from '@mui/material';
 import LogoutButton from '../atoms/LogoutButton';
 import PetStatusModal from './petStatusModal';
 import { getPetDetails } from '../../features/api/getPetDetails';
@@ -13,8 +11,7 @@ import { getPetInfo } from '../../features/api/getPetInfo';
 import { AnchorTemporaryDrawerProps, Anchor, PetDetails } from '../../types/index'
 
 import { mainTheme } from '../../styles/theme'
-
-const cache = createCache({ key: 'css', prepend: true });
+import ThemeWrapper from '../../styles/ThemeWrapper';
 
 export default function AnchorTemporaryDrawer({ petDetails, setPetDetails, onFeed, onStroke, onPlay, setOffspringCount, physicalRecoveryIntervalId, statDecreaseIntervalId }: AnchorTemporaryDrawerProps & { 
   petDetails: PetDetails | null; 
@@ -166,33 +163,30 @@ export default function AnchorTemporaryDrawer({ petDetails, setPetDetails, onFee
     );
 
   return(
-    <CacheProvider value={cache}>
-      <ThemeProvider theme={mainTheme}>
-        <CssBaseline />
-            <React.Fragment>
-              <Button
-                onClick={toggleDrawer("top", true)}
-                variant="contained"
-                color="secondary"
-                sx={{ color: 'white', marginTop:"7vh", fontWeight:"bold", fontSize:"24px", zIndex: 1000  }}>
-                  menu
-              </Button>
-              <Drawer
-                anchor="top"
-                open={state["top"]}
-                onClose={toggleDrawer("top", false)}
-                sx={{ zIndex: 1000 }}
-              >
-                {list("top")}
-              </Drawer>
-              <PetStatusModal
-                open={openModal}
-                onClose={handleCloseModal}
-                petDetails={petDetails}
-              />
-            </React.Fragment>
-      </ThemeProvider>
-    </CacheProvider>
+    <ThemeWrapper theme={mainTheme}>
+      <React.Fragment>
+        <Button
+          onClick={toggleDrawer("top", true)}
+          variant="contained"
+          color="secondary"
+          sx={{ color: 'white', marginTop:"7vh", fontWeight:"bold", fontSize:"24px", zIndex: 1000  }}>
+            menu
+        </Button>
+        <Drawer
+          anchor="top"
+          open={state["top"]}
+          onClose={toggleDrawer("top", false)}
+          sx={{ zIndex: 1000 }}
+        >
+          {list("top")}
+        </Drawer>
+        <PetStatusModal
+          open={openModal}
+          onClose={handleCloseModal}
+          petDetails={petDetails}
+        />
+      </React.Fragment>
+    </ThemeWrapper>
   )
 
 }
