+++ styles/ThemeWrapper.tsx
@@ -1,0 +1,24 @@
+import React, { ReactNode } from 'react';
+import { CacheProvider } from '@emotion/react';
+import createCache from '@emotion/cache';
+import { CssBaseline, ThemeProvider, Theme } from '@mui/material';
+
+const cache = createCache({ key: 'css', prepend: true });
+
+interface ThemeWrapperProps {
+  theme: Theme;
+  children: ReactNode;
+}
+
+const ThemeWrapper: React.FC<ThemeWrapperProps> = ({ theme, children }) => {
+  return (
+    <CacheProvider value={cache}>
+      <ThemeProvider theme={theme}>
+        <CssBaseline />
+        {children}
+      </ThemeProvider>
+    </CacheProvider>
+  );
+};
+
+export default ThemeWrapper;
